refactor: move executor tests
