+++ tests/executor/indexer/elastic/test_executor.py
@@ -1,222 +1,266 @@
-from typing import List, Union
-
-import pytest
-from docarray import DocumentArray
-from elasticsearch import Elasticsearch
-from jina import Flow
-
-from now.constants import ModelDimensions
-from now.executor.indexer.elastic import NOWElasticIndexer
-
-
-@pytest.mark.skip(reason="executor currently not working and will be updated soon")
-@pytest.mark.parametrize(
-    'da, dims',
-    [('text_da', 7), ('multimodal_da', [7, 5])],
+from now.executor.indexer.elastic.elastic_indexer import (
+    FieldEmbedding,
+    NOWElasticIndexer,
 )
-def test_indexing(
-    da: DocumentArray,
-    dims: Union[str, List[int]],
-    setup_service_running,
-    es_connection_params,
-    request,
-):
-    da = request.getfixturevalue(da)
-    index_name = 'test-indexing'
-    hosts, _ = es_connection_params
-    with Flow().add(
-        uses=NOWElasticIndexer,
-        uses_with={
-            'hosts': hosts,
-            'index_name': index_name,
-            'dims': dims,
-        },
-    ) as f:
-        f.index(da)
-        es = Elasticsearch(hosts=hosts)
-        # fetch all indexed documents
-        res = es.search(index=index_name, size=100, query={'match_all': {}})
-        assert len(res['hits']['hits']) == 2
 
 
-@pytest.mark.skip(reason="executor currently not working and will be updated soon")
-@pytest.mark.parametrize(
-    'da, query_da, dims, index_name, apply_bm25',
-    [
-        ('text_da', 'text_query', 7, 'test-search', False),
-        (
-            'multimodal_da',
-            'multimodal_query',
-            [7, 5],
-            'test-search-multimodal',
-            False,
-        ),
-        ('text_da', 'text_query', 7, 'test-search-bm25-text', True),
-        (
-            'multimodal_da',
-            'multimodal_query',
-            [7, 5],
-            'test-search-bm25-multimodal',
-            True,
-        ),
-    ],
-)
-def test_search_with_bm25(
-    da: DocumentArray,
-    query_da: DocumentArray,
-    dims: Union[int, List[int]],
-    index_name: str,
-    apply_bm25: bool,
-    setup_service_running,
-    es_connection_params,
-    request,
-):
-    da = request.getfixturevalue(da)
-    query_da = request.getfixturevalue(query_da)
-    hosts, _ = es_connection_params
-    with Flow().add(
-        uses=NOWElasticIndexer,
-        uses_with={
-            'hosts': hosts,
-            'index_name': index_name,
-            'dims': dims,
-        },
-    ) as f:
-        f.index(da)
-        x = f.search(
-            query_da,
-            parameters={'apply_bm25': apply_bm25},
-        )
-        assert len(x) != 0
-        assert len(x[0].matches) != 0
-
-
-@pytest.mark.skip(reason="executor currently not working and will be updated soon")
-@pytest.mark.parametrize(
-    'da, query_da, dims, index_name',
-    [
-        ('text_da', 'text_query', 7, 'test-search-filter-text'),
-        (
-            'multimodal_da',
-            'multimodal_query',
-            [7, 5],
-            'test-search-filter-multimodal',
-        ),
-    ],
-)
-def test_search_with_filter(
-    da: DocumentArray,
-    query_da: DocumentArray,
-    index_name: str,
-    dims: Union[int, List[int]],
-    setup_service_running,
-    es_connection_params,
-    request,
-):
-    da = request.getfixturevalue(da)
-    query_da = request.getfixturevalue(query_da)
-    hosts, _ = es_connection_params
-    with Flow().add(
-        uses=NOWElasticIndexer,
-        uses_with={
-            'hosts': hosts,
-            'index_name': index_name,
-            'dims': dims,
-        },
-    ) as f:
-        f.index(da)
-        x = f.search(
-            query_da,
-            parameters={'apply_bm25': True, 'filter': {'cost': {'gte': 20.0}}},
-        )
-        assert len(x) != 0
-        assert len(x[0].matches) == 1
-
-
-@pytest.mark.skip(reason="executor currently not working and will be updated soon")
-@pytest.mark.parametrize(
-    'da, dims, index_name',
-    [
-        ('text_da', 7, 'test-list-text'),
-        ('multimodal_da', [7, 5], 'test-list-multimodal'),
-    ],
-)
-def test_list(
-    da: DocumentArray,
-    dims: Union[int, List[int]],
-    index_name: str,
-    setup_service_running,
-    es_connection_params,
-    request,
-):
-    da = request.getfixturevalue(da)
-    hosts, _ = es_connection_params
-    with Flow().add(
-        uses=NOWElasticIndexer,
-        uses_with={
-            'hosts': hosts,
-            'index_name': index_name,
-            'dims': dims,
-        },
-    ) as f:
-        f.index(da)
-        list_res = f.post(on='/list', parameters={'limit': 1, 'offset': 1})
-        assert list_res[0].id == '456'
-
-
-@pytest.mark.skip(reason="executor currently not working and will be updated soon")
-@pytest.mark.parametrize(
-    'da, dims, index_name',
-    [
-        ('text_da', 7, 'test-delete-text'),
-        ('multimodal_da', [7, 5], 'test-delete-multimodal'),
-    ],
-)
-def test_delete(
-    da: DocumentArray,
-    dims: Union[int, List[int]],
-    index_name: str,
-    setup_service_running,
-    es_connection_params,
-    request,
-):
-    da = request.getfixturevalue(da)
-    hosts, _ = es_connection_params
-    with Flow().add(
-        uses=NOWElasticIndexer,
-        uses_with={
-            'hosts': hosts,
-            'index_name': index_name,
-            'dims': dims,
-        },
-    ) as f:
-        f.index(da)
-        es = Elasticsearch(hosts=hosts)
-        # fetch all indexed documents
-        res = es.search(index=index_name, size=100, query={'match_all': {}})
-        assert len(res['hits']['hits']) == 2
-        f.post(on='/delete', parameters={'filter': {'cost': {'gte': 15.0}}})
-        res = es.search(index=index_name, size=100, query={'match_all': {}})
-        assert len(res['hits']['hits']) == 0
-
-
-@pytest.mark.skip(reason="executor currently not working and will be updated soon")
-@pytest.mark.parametrize(
-    'docs_matrix, on',
-    [
-        ('docs_matrix_index', 'index'),
-        ('docs_matrix_search', 'search'),
-    ],
-)
-def test_merge_docs_matrix(
-    docs_matrix: List[DocumentArray],
-    on: str,
-    request,
-):
-    docs_matrix = request.getfixturevalue(docs_matrix)
-    merged_result = NOWElasticIndexer._join_docs_matrix_into_chunks(
-        on=on, docs_matrix=docs_matrix
+def test_generate_es_mappings(setup_service_running):
+    """
+    This test should check, whether the static
+    generate_es_mappings method works as expected.
+    """
+    document_mappings = [
+        FieldEmbedding('clip', 8, ['title']),
+        FieldEmbedding('sbert', 5, ['title', 'excerpt']),
+    ]
+    expected_mapping = {
+        'properties': {
+            'id': {'type': 'keyword'},
+            'bm25_text': {'type': 'text', 'analyzer': 'standard'},
+            'title-clip': {
+                'properties': {
+                    'embedding': {
+                        'type': 'dense_vector',
+                        'dims': '8',
+                        'similarity': 'cosine',
+                        'index': 'true',
+                    }
+                }
+            },
+            'title-sbert': {
+                'properties': {
+                    'embedding': {
+                        'type': 'dense_vector',
+                        'dims': '5',
+                        'similarity': 'cosine',
+                        'index': 'true',
+                    },
+                }
+            },
+            'excerpt-sbert': {
+                'properties': {
+                    'embedding': {
+                        'type': 'dense_vector',
+                        'dims': '5',
+                        'similarity': 'cosine',
+                        'index': 'true',
+                    }
+                }
+            },
+        }
+    }
+    result = NOWElasticIndexer.generate_es_mapping(
+        document_mappings=document_mappings, metric='cosine'
     )
-    assert len(merged_result[0].chunks) == 2
-    assert merged_result[0].chunks[0].embedding.shape == (ModelDimensions.SBERT,)
-    assert merged_result[0].chunks[1].embedding.shape == (ModelDimensions.CLIP,)
+    assert result == expected_mapping
+
+
+def test_index_and_search_with_multimodal_docs(
+    setup_service_running, es_inputs, random_index_name
+):
+    """
+    This test runs indexing with the NOWElasticIndexer using multimodal docs.
+    """
+    (
+        index_docs_map,
+        query_docs_map,
+        document_mappings,
+        default_semantic_scores,
+    ) = es_inputs
+
+    indexer = NOWElasticIndexer(
+        traversal_paths='c',
+        document_mappings=document_mappings,
+        default_semantic_scores=default_semantic_scores,
+        # es_config={'api_key': os.environ['ELASTIC_API_KEY']},
+        # hosts='https://5280f8303ccc410295d02bbb1f3726f7.eu-central-1.aws.cloud.es.io:443',
+        hosts='http://localhost:9200',
+        index_name=random_index_name,
+        document_structure='MMDoc()',
+    )
+
+    indexer.index(index_docs_map)
+    # check if documents are indexed
+    es = indexer.es
+    res = es.search(index=random_index_name, size=100, query={'match_all': {}})
+    assert len(res['hits']['hits']) == len(index_docs_map['clip'])
+    results = indexer.search(
+        query_docs_map,
+        parameters={'get_score_breakdown': True, 'apply_default_bm25': True},
+    )
+    print(results[0].matches[0].scores)
+    # asserts about matches
+    for (
+        query_field,
+        document_field,
+        encoder,
+        linear_weight,
+    ) in default_semantic_scores:
+        if encoder == 'bm25':
+            assert 'bm25_normalized' in results[0].matches[0].scores
+            assert 'bm25_raw' in results[0].matches[0].scores
+            assert isinstance(
+                results[0].matches[0].scores['bm25_normalized'].value, float
+            )
+            assert isinstance(results[0].matches[0].scores['bm25_raw'].value, float)
+        else:
+            score_string = '-'.join(
+                [
+                    query_field,
+                    document_field,
+                    encoder,
+                    str(linear_weight),
+                ]
+            )
+            assert score_string in results[0].matches[0].scores
+            assert isinstance(results[0].matches[0].scores[score_string].value, float)
+
+
+def test_list_endpoint(setup_service_running, es_inputs, random_index_name):
+    """
+    This test tests the list endpoint of the NOWElasticIndexer.
+    """
+    (
+        index_docs_map,
+        query_docs_map,
+        document_mappings,
+        default_semantic_scores,
+    ) = es_inputs
+    es_indexer = NOWElasticIndexer(
+        traversal_paths='c',
+        document_mappings=document_mappings,
+        default_semantic_scores=default_semantic_scores,
+        hosts='http://localhost:9200',
+        index_name=random_index_name,
+    )
+    es_indexer.index(index_docs_map)
+    result = es_indexer.list()
+    assert len(result) == len(index_docs_map['clip'])
+    limit = 1
+    result_with_limit = es_indexer.list(parameters={'limit': limit})
+    assert len(result_with_limit) == limit
+    offset = 1
+    result_with_offset = es_indexer.list(parameters={'offset': offset})
+    assert len(result_with_offset) == len(index_docs_map['clip']) - offset
+
+
+def test_delete_by_id(setup_service_running, es_inputs, random_index_name):
+    """
+    This test tests the delete endpoint of the NOWElasticIndexer, by deleting a list of IDs.
+    """
+    (
+        index_docs_map,
+        query_docs_map,
+        document_mappings,
+        default_semantic_scores,
+    ) = es_inputs
+    es_indexer = NOWElasticIndexer(
+        traversal_paths='c',
+        document_mappings=document_mappings,
+        default_semantic_scores=default_semantic_scores,
+        hosts='http://localhost:9200',
+        index_name=random_index_name,
+    )
+    es_indexer.index(index_docs_map)
+    # delete by id
+    ids = [doc.id for doc in index_docs_map['clip']]
+    es_indexer.delete(parameters={'ids': ids})
+
+    es = es_indexer.es
+    res = es.search(index=random_index_name, size=100, query={'match_all': {}})
+    assert len(res['hits']['hits']) == 0
+
+
+def test_custom_mapping_and_custom_bm25_search(
+    setup_service_running, es_inputs, random_index_name
+):
+    """
+    This test tests the custom mapping and bm25 functionality of the NOWElasticIndexer.
+    """
+    (
+        index_docs_map,
+        query_docs_map,
+        document_mappings,
+        default_semantic_scores,
+    ) = es_inputs
+    es_mapping = {
+        'properties': {
+            'id': {'type': 'keyword'},
+            'bm25_text': {'type': 'text', 'analyzer': 'standard'},
+            'title-clip': {
+                'properties': {
+                    'embedding': {
+                        'type': 'dense_vector',
+                        'dims': '8',
+                        'similarity': 'cosine',
+                        'index': 'true',
+                    }
+                }
+            },
+            'title-sbert': {
+                'properties': {
+                    'embedding': {
+                        'type': 'dense_vector',
+                        'dims': '5',
+                        'similarity': 'cosine',
+                        'index': 'true',
+                    }
+                }
+            },
+            'gif-clip': {
+                'properties': {
+                    'embedding': {
+                        'type': 'dense_vector',
+                        'dims': '8',
+                        'similarity': 'cosine',
+                        'index': 'true',
+                    }
+                }
+            },
+            'excerpt-sbert': {
+                'properties': {
+                    'embedding': {
+                        'type': 'dense_vector',
+                        'dims': '5',
+                        'similarity': 'cosine',
+                        'index': 'true',
+                    }
+                }
+            },
+        }
+    }
+    es_indexer = NOWElasticIndexer(
+        traversal_paths='c',
+        document_mappings=document_mappings,
+        default_semantic_scores=default_semantic_scores,
+        es_mapping=es_mapping,
+        hosts='http://localhost:9200',
+        index_name=random_index_name,
+    )
+    # do indexing
+    es_indexer.index(index_docs_map)
+    # search with custom bm25 query with field boosting
+    custom_bm25_query = {
+        'multi_match': {
+            'query': 'this cat is cute',
+            'fields': ['bm25_text^7'],
+            'tie_breaker': 0.3,
+        }
+    }
+    results = es_indexer.search(
+        query_docs_map,
+        parameters={
+            'get_score_breakdown': True,
+            'custom_bm25_query': custom_bm25_query,
+        },
+    )
+    assert len(results[0].matches) == 2
+    assert results[0].matches[0].id == '0'
+    assert results[0].matches[1].id == '1'
+
+
+def test_search_with_filter(setup_service_running, es_inputs, random_index_name):
+    """
+    TODO: fill
+    """
+    pass
